corrected security and auth-shaking in backend

+++ api-commons/src/main/java/edu/usach/apicommons/controller/EntityController.java
@@ -1,21 +1,18 @@
 package edu.usach.apicommons.controller;
 
 import edu.usach.apicommons.errorhandling.ApiException;
-import edu.usach.apicommons.errorhandling.ErrorDTO;
+import edu.usach.apicommons.errorhandling.ErrorCode;
 import edu.usach.apicommons.model.IEntity;
-import edu.usach.apicommons.model.ISecureEntity;
 import edu.usach.apicommons.service.IEntityService;
-import edu.usach.apicommons.util.SecurityUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.json.simple.JSONObject;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.lang.reflect.ParameterizedType;
 
-import static edu.usach.apicommons.util.Constants.ARRAY;
 import static edu.usach.apicommons.util.Constants.OBJECT;
-import static edu.usach.apicommons.util.SecurityUtils.HEADER_STRING;
 
 @SuppressWarnings("unchecked")
 @Slf4j
@@ -23,33 +20,23 @@ public abstract class EntityController<T extends IEntity> extends AbstractContro
 
 	private Class<T> tClass = (Class<T>) ((ParameterizedType) getClass().getGenericSuperclass()).getActualTypeArguments()[0];
 
-	protected boolean isAuthenticated() {
-		return SecurityUtils.isAuthenticated(httpServletRequest.getHeader(HEADER_STRING));
-	}
-
-	protected boolean isAuthorized(String... roles) {
-		return SecurityUtils.isAuthorized(httpServletRequest.getHeader(HEADER_STRING), roles);
-	}
-
-	protected boolean hasAccess(ISecureEntity entity) {
-		return SecurityUtils.hasAccess(httpServletRequest.getHeader(HEADER_STRING), entity);
-	}
-
 	protected abstract IEntityService<T> getService();
 
 	@Override
 	@RequestMapping(method = RequestMethod.GET, value = "/{id}")
 	public ResponseEntity<Object> getById(@PathVariable Long id, @RequestParam(value = "show", required = false) String filterString) {
+		if (!isAuthenticated())
+			return responseApiException(new ApiException(ErrorCode.UNAUTHORIZED), HttpStatus.UNAUTHORIZED);
 		try {
 			if (null == filterString)
 				return response(getService().findOne(id));
 			return response(getService().findAndFilter(id, filterString));
 		} catch (ApiException e) {
 			log.error(e.getMessage(), e);
-			return responseNotFound(OBJECT, new ErrorDTO(e, httpServletRequest));
+			return responseApiException(e, HttpStatus.NOT_FOUND);
 		} catch (Exception e) {
 			log.error(e.getMessage(), e);
-			return responseInternalServerError(OBJECT, new ErrorDTO(httpServletRequest));
+			return responseException(e, HttpStatus.INTERNAL_SERVER_ERROR);
 		}
 	}
 
@@ -58,41 +45,48 @@ public abstract class EntityController<T extends IEntity> extends AbstractContro
 	public ResponseEntity<Object> getAll(
 			@RequestParam(value = "page", required = false) Integer page,
 			@RequestParam(value = "size", required = false) Integer size,
-			@RequestParam(value = "showAll", required = false) String showAll) {
+			@RequestParam(value = "showAll", required = false) String showAll
+	) {
+		if (!isAuthenticated())
+			return responseApiException(new ApiException(ErrorCode.UNAUTHORIZED), HttpStatus.UNAUTHORIZED);
 		try {
 			if (null!=showAll && showAll.matches("(?i)^(yes|true|ok|1|show|showall)$"))
 				return response(getService().findAll());
 			return response(getService().findPaginated(page, size));
 		} catch (ApiException e) {
 			log.error(e.getMessage(), e);
-			return responseNotFound(ARRAY, new ErrorDTO(e, httpServletRequest));
+			return responseApiException(e, HttpStatus.NOT_FOUND);
 		} catch (Exception e) {
 			log.error(e.getMessage(), e);
-			return responseInternalServerError(ARRAY, new ErrorDTO(httpServletRequest));
+			return responseException(e, HttpStatus.INTERNAL_SERVER_ERROR);
 		}
 	}
 
 	@Override
 	@RequestMapping(method = RequestMethod.POST)
 	public ResponseEntity<Object> create(@RequestBody T entity) {
+		if (!isAuthenticated())
+			return responseApiException(new ApiException(ErrorCode.UNAUTHORIZED), HttpStatus.UNAUTHORIZED);
 		try {
 			getService().create(entity);
 			JSONObject data = new JSONObject();
 			data.put("success", true);
 			data.put("message", tClass.getName() + " successfully created");
-			return responseCreated(data);
+			return response(data, HttpStatus.CREATED);
 		} catch (ApiException e) {
 			log.error(e.getMessage(), e);
-			return responseNotFound(OBJECT, new ErrorDTO(e, httpServletRequest));
+			return responseApiException(e, HttpStatus.NOT_FOUND);
 		} catch (Exception e) {
 			log.error(e.getMessage(), e);
-			return responseInternalServerError(OBJECT, new ErrorDTO(httpServletRequest));
+			return responseException(e, HttpStatus.INTERNAL_SERVER_ERROR);
 		}
 	}
 
 	@Override
 	@RequestMapping(method = RequestMethod.PUT)
 	public ResponseEntity<Object> update(@RequestBody T entity) {
+		if (!isAuthenticated())
+			return responseApiException(new ApiException(ErrorCode.UNAUTHORIZED), HttpStatus.UNAUTHORIZED);
 		try {
 			getService().update(entity);
 			JSONObject data = new JSONObject();
@@ -101,23 +95,25 @@ public abstract class EntityController<T extends IEntity> extends AbstractContro
 			return response(data);
 		} catch (ApiException e) {
 			log.error(e.getMessage(), e);
-			return responseNotFound(OBJECT, new ErrorDTO(e, httpServletRequest));
+			return responseApiException(e, HttpStatus.NOT_FOUND);
 		} catch (Exception e) {
 			log.error(e.getMessage(), e);
-			return responseInternalServerError(OBJECT, new ErrorDTO(httpServletRequest));
+			return responseException(e, HttpStatus.INTERNAL_SERVER_ERROR);
 		}
 	}
 
 	@Override
 	@RequestMapping(method = RequestMethod.DELETE)
 	public ResponseEntity<Object> delete(@RequestBody(required = false) T entity, @RequestParam(value = "id", required = false) Long id) {
+		if (!isAuthenticated())
+			return responseApiException(new ApiException(ErrorCode.UNAUTHORIZED), HttpStatus.UNAUTHORIZED);
 		try {
 			if (null!=id && null==entity) {
 				getService().deleteById(id);
 			}else if (null!=entity && null==id) {
 				getService().delete(entity);
 			} else {
-				return responseBadRequest(OBJECT, new ErrorDTO(httpServletRequest));
+				return response(OBJECT, HttpStatus.BAD_REQUEST);
 			}
 			JSONObject data = new JSONObject();
 			data.put("success", true);
@@ -125,10 +121,10 @@ public abstract class EntityController<T extends IEntity> extends AbstractContro
 			return response(data);
 		} catch (ApiException e) {
 			log.error(e.getMessage(), e);
-			return responseNotFound(OBJECT, new ErrorDTO(e, httpServletRequest));
+			return responseApiException(e, HttpStatus.NOT_FOUND);
 		} catch (Exception e) {
 			log.error(e.getMessage(), e);
-			return responseInternalServerError(OBJECT, new ErrorDTO(httpServletRequest));
+			return responseException(e, HttpStatus.INTERNAL_SERVER_ERROR);
 		}
 	}
